Fixed resetting of GeoJSONTileLayer
Some problems occur when the a category was changed.
This leads to a few things:
  - zoom level change, plenty of new tiles are requested
  - url change
  - cancelling of all ongoing tile requests

When the url is changed, there might still be new requests
queued with the old url, added check to prevent that new
results are added.

Aborting was ignored on some platforms, added local flag.
Ensure tiles are cleaned up correctly.