Remove multiple viewports meta tags
This is all a very simple fix for a kindof complicated situation.
next/head will automatically add two default metas, a charset & a viewport one.

The viewport meta  does not contain any viewport-fit=cover.
If the meta is added in addition to the one from the app, the last added one wins,
this breaks behavior in the native app under ios on an iphone x.

To fix this we move the meta header into a next/head so that the deduplication will
work. We cannot rely on deduplication by key as mentioned in the next documentation,
as React.mapChildren will prefix keys with ".$", but the keys of the default-metas
in next/head are not prefixed. Deduplication by name/meta-type works fine.