Reload equipment layer once every hour
This solves a few of the problems with outdated equipment infos
being displayed in the app when returning to an old session.

Instead of never touching the map tiles again, the map tiles are
reloaded after half the time it takes to become outdated. As the
information updates every 10minutes on ac, the current 60 minutes
should be fine.

This is cheaper than the opposite approach of reloading all the
elements in the EquipmentInfoCache, as only the visible tiles will
be loaded.

This change does not fix the issue that the place details stay
unrefreshed until a change is triggered on the site. This requires
another timer in the place details, or a change to the feature cache
as so far events are not fired for map tile loads.