Simplify toolbar positioning
Data should always flow down and positioning should always happen in parents, not in children. This adds a `minimalTopPosition` property to the Toolbar component, allowing the caller to customize the toolbar’s minimal top position.

This allows the MainView, which knows about the relationship between <iframe> embedding and toolbar positioning, to position the toolbar correctly.