Fix missing map marker issue
This happened because the tile URL was generated randomly if multi-hostname tile source URLs were used.

Now, the tile URLs are generated not randomly anymore, but in a round-robin scheme mathematically tied to the x coordinate of the tile. This won't work as well in fully zoomed out modes in theory, but OSM places from tiles are only shown in zoomed in views.

This also saves CDN caching storage as each of the 4 CDN varnish servers only acts as a cache for a fourth of the tiles.